More WIP on separation between the MooseHelper and the GUI/settings app. Currently, settings are not yet shared, meaning the settings app is "broken".

+++ MooseBetterLipSync/UKMouthShape.h
@@ -31,7 +31,7 @@ enum
 @interface UKMouthShape : NSObject <NSCopying>
 {
 	UKMouthPoint*	points;
-	int				numPoints;
+	NSUInteger		numPoints;
 	NSSize			imageSize;
 }
